+++ nobodywho/core/src/lib.rs
@@ -8,6 +8,7 @@ mod host_memory;
 pub mod huggingface;
 pub mod inference;
 pub mod llm;
+pub mod logging;
 pub mod memory;
 mod model_selection;
 pub mod onnx;